Refactor availablePaths and libraryCategories into shared constant
