+++ kolibri/plugins/learn/assets/src/views/CategorySearchModal/CategorySearchOptions.vue
@@ -48,51 +48,8 @@
 
   import camelCase from 'lodash/camelCase';
   import commonCoreStrings from 'kolibri.coreVue.mixins.commonCoreStrings';
-  import { Categories, CategoriesLookup } from 'kolibri.coreVue.vuex.constants';
   import responsiveWindowMixin from 'kolibri.coreVue.mixins.responsiveWindowMixin';
-  import plugin_data from 'plugin_data';
-
-  const availablePaths = {};
-
-  if (process.env.NODE_ENV !== 'production') {
-    // TODO rtibbles: remove this condition
-    Object.assign(availablePaths, CategoriesLookup);
-  } else {
-    plugin_data.categories.map(key => {
-      const paths = key.split('.');
-      let path = '';
-      for (let path_segment of paths) {
-        path = path === '' ? path_segment : path + '.' + path_segment;
-        availablePaths[path] = true;
-      }
-    });
-  }
-
-  const libraryCategories = {};
-
-  for (let subjectKey of Object.entries(Categories)
-    .sort((a, b) => a[0].length - b[0].length)
-    .map(a => a[0])) {
-    const ids = Categories[subjectKey].split('.');
-    let path = '';
-    let nested = libraryCategories;
-    for (let fragment of ids) {
-      path += fragment;
-      if (availablePaths[path]) {
-        const nestedKey = CategoriesLookup[path];
-        if (!nested[nestedKey]) {
-          nested[nestedKey] = {
-            value: path,
-            nested: {},
-          };
-        }
-        nested = nested[nestedKey].nested;
-        path += '.';
-      } else {
-        break;
-      }
-    }
-  }
+  import { libraryCategories } from '../../constants';
 
   export default {
     name: 'CategorySearchOptions',
